chore(gapic-common): Revert gapic polling harness
This reverts commit ab832c199f1364bd5bbff09396bbb6d77b663e10.
